保存一个里程碑

+++ order-common/src/main/java/com/erui/order/common/enums/DeliverNoticeStatusEnum.java
@@ -15,7 +15,7 @@ public enum DeliverNoticeStatusEnum {
         this.name = name;
     }
 
-    public int getCode() {
+    public short getCode() {
         return code;
     }
