+++ order-common/src/main/java/com/erui/order/common/enums/InstockStatusEnum.java
@@ -15,7 +15,7 @@ public enum InstockStatusEnum {
         this.name = name;
     }
 
-    public int getCode() {
+    public short getCode() {
         return code;
     }
